move private structs into source files, ensure private functions are static in nvme.c
