+++ kernel/dev/storage/nvme/nvme.c
@@ -8,6 +8,257 @@
 #include <lib/resource.h>
 #include <mm/vmm.h>
 
+// capabilities for the nvme controller
+struct nvme_id {
+    uint16_t vid;
+    uint16_t ssvid;
+    char sn[20];
+    char mn[40];
+    char fr[8];
+    uint8_t rab;
+    uint8_t ieee[3];
+    uint8_t mic;
+    uint8_t mdts;
+    uint16_t ctrlid;
+    uint32_t version;
+    uint32_t unused1[43];
+    uint16_t oacs;
+    uint8_t acl;
+    uint8_t aerl;
+    uint8_t fw; // firmware
+    uint8_t lpa;
+    uint8_t elpe;
+    uint8_t npss;
+    uint8_t avscc;
+    uint8_t apsta;
+    uint16_t wctemp;
+    uint16_t cctemp;
+    uint16_t unused2[121];
+    uint8_t sqes;
+    uint8_t cqes;
+    uint16_t unused3;
+    uint32_t nn;
+    uint16_t oncs;
+    uint16_t fuses;
+    uint8_t fna;
+    uint8_t vwc;
+    uint16_t awun;
+    uint16_t awupf;
+    uint8_t nvscc;
+    uint8_t unused4;
+    uint16_t acwu;
+    uint16_t unused5;
+    uint32_t sgls;
+    uint32_t unused6[1401];
+    uint8_t vs[1024];
+};
+
+struct nvme_lbaf {
+    uint16_t ms;
+    uint8_t ds;
+    uint8_t rp;
+};
+
+struct nvme_nsid {
+    uint64_t size;
+    uint64_t capabilities;
+    uint64_t nuse;
+    uint8_t features;
+    uint8_t nlbaf;
+    uint8_t flbas;
+    uint8_t mc;
+    uint8_t dpc;
+    uint8_t dps;
+    uint8_t nmic;
+    uint8_t rescap;
+    uint8_t fpi;
+    uint8_t unused1;
+    uint16_t nawun;
+    uint16_t nawupf;
+    uint16_t nacwu;
+    uint16_t nabsn;
+    uint16_t nabo;
+    uint16_t nabspf;
+    uint16_t unused2;
+    uint64_t nvmcap[2];
+    uint64_t unusued3[5];
+    uint8_t nguid[16];
+    uint8_t eui64[8];
+    struct nvme_lbaf lbaf[16];
+    uint64_t unused3[24];
+    uint8_t vs[3712];
+};
+
+// generic
+#define NVME_OPFLUSH 0x00
+#define NVME_OPWRITE 0x01
+#define NVME_OPREAD 0x02
+// admin
+#define NVME_OPCREATESQ 0x01
+#define NVME_OPDELCQ 0x04
+#define NVME_OPCREATECQ 0x05
+#define NVME_OPIDENTIFY 0x06
+#define NVME_OPABORT 0x08
+#define NVME_OPSETFT 0x09
+#define NVME_OPGETFT 0x0a
+
+// each command is 512 bytes, so we need to use a union along with reserving unused data portions
+// somewhat complete NVMe command set
+struct nvme_cmd {
+    union {
+        struct {
+            uint8_t opcode; 
+            uint8_t flags; 
+            uint16_t cid; 
+            uint32_t nsid; 
+            uint32_t cdw1[2]; 
+            uint64_t metadata; 
+            uint64_t prp1; 
+            uint64_t prp2; 
+            uint32_t cdw2[6];
+        } common; // generic command
+        struct {
+            uint8_t opcode; 
+            uint8_t flags; 
+            uint16_t cid; 
+            uint32_t nsid; 
+            uint64_t unused;
+            uint64_t metadata; 
+            uint64_t prp1; 
+            uint64_t prp2; 
+            uint64_t slba; 
+            uint16_t len; 
+            uint16_t control; 
+            uint32_t dsmgmt; 
+            uint32_t ref; 
+            uint16_t apptag; 
+            uint16_t appmask;
+        } rw; // read or write
+        struct {
+            uint8_t opcode; 
+            uint8_t flags; 
+            uint16_t cid; 
+            uint32_t nsid; 
+            uint64_t unused1; 
+            uint64_t unused2; 
+            uint64_t prp1; 
+            uint64_t prp2; 
+            uint32_t cns; 
+            uint32_t unused3[5]; 
+        } identify; // identify
+        struct { uint8_t opcode; 
+            uint8_t flags; 
+            uint16_t cid;
+            uint32_t nsid; 
+            uint64_t unused1;
+            uint64_t unused2; 
+            uint64_t prp1; 
+            uint64_t prp2; 
+            uint32_t fid; 
+            uint32_t dword;
+            uint64_t unused[2]; 
+        } features;
+        struct { uint8_t opcode; 
+            uint8_t flags; 
+            uint16_t cid; 
+            uint32_t unused1[5]; 
+            uint64_t prp1;
+            uint64_t unused2; 
+            uint16_t cqid; 
+            uint16_t size; 
+            uint16_t cqflags; 
+            uint16_t irqvec; 
+            uint64_t unused3[2];
+        } createcompq;
+        struct { uint8_t opcode;
+            uint8_t flags; 
+            uint16_t cid; 
+            uint32_t unused1[5];
+            uint64_t prp1; 
+            uint64_t unused2; 
+            uint16_t sqid; 
+            uint16_t size; 
+            uint16_t sqflags; 
+            uint16_t cqid; 
+            uint64_t unused3[2]; 
+        } createsubq;
+        struct { uint8_t opcode; 
+            uint8_t flags;
+            uint16_t cid;
+            uint32_t unused1[9];
+            uint16_t qid;
+            uint16_t unused2; 
+            uint32_t unused3[5]; 
+        } deleteq;
+        struct { uint8_t opcode;
+            uint8_t flags; 
+            uint16_t cid;
+            uint32_t unused1[9];
+            uint16_t sqid;
+            uint16_t cqid; 
+            uint32_t unused2[5]; 
+        } abort;
+    };
+};
+
+// command result
+struct nvme_cmdcomp {
+    uint32_t res;
+    uint32_t unused;
+    uint16_t sqhead;
+    uint16_t sqid;
+    uint16_t cid;
+    uint16_t status;
+};
+
+// according to BAR0 registers (osdev wiki)
+struct nvme_bar {
+    uint64_t capabilities; // 0x00-0x07
+    uint32_t version; // 0x08-0x0B
+    uint32_t intms; // 0x0C-0x0F (interrupt mask set)
+    uint32_t intmc; // 0x10-0x13 (interrupt mask clear)
+    uint32_t conf; // 0x14-0x17
+    uint32_t unused1;
+    uint32_t status; // 0x1C-0x1F
+    uint32_t unused2;
+    uint32_t aqa; // 0x24-0x27 (admin queue attrs)
+    uint64_t asq; // 0x28-0x2F (admin submit queue)
+    uint64_t acq; // 0x30-0x37 (admin completion queue)
+};
+
+#define NVME_CAPMQES(cap)      ((cap) & 0xffff)
+#define NVME_CAPSTRIDE(cap)    (((cap) >> 32) & 0xf)
+#define NVME_CAPMPSMIN(cap)    (((cap) >> 48) & 0xf)
+#define NVME_CAPMPSMAX(cap)    (((cap) >> 52) & 0xf)
+
+struct nvme_queue {
+    volatile struct nvme_cmd *submit;
+    volatile struct nvme_cmdcomp *completion;
+    volatile uint32_t *submitdb;  
+    volatile uint32_t *completedb;
+    uint16_t elements; // elements in queue
+    uint16_t cqvec;
+    uint16_t sqhead;
+    uint16_t sqtail;
+    uint16_t cqhead;
+    uint8_t cqphase;
+    uint16_t qid; // queue id
+    uint32_t cmdid; // command id
+    uint64_t *physregpgs; // pointer to the PRPs
+};
+
+#define NVME_WAITCACHE 0 // cache is not ready
+#define NVME_READYCACHE 1 // cache is ready
+#define NVME_DIRTYCACHE 2 // cache is damaged/dirty
+
+// cache (helps speed up disk reads by hitting cache)
+struct cachedblock {
+    uint8_t *cache; // pointer to the cache we have for this block
+    uint64_t block;
+    uint64_t end;
+    int status;
+};
+
 // controller
 struct nvme_device {
     struct resource;
@@ -67,6 +318,47 @@ static void nvme_createqueue(struct nvme_nsdevice *ns, struct nvme_queue *queue,
     queue->physregpgs = alloc(ns->maxphysrpgs * slots * sizeof(uint64_t));
 }
 
+// submit a command
+static void nvme_submitcmd(struct nvme_queue *queue, struct nvme_cmd cmd) {
+    uint16_t tail = queue->sqtail; // tail of the submit queue
+    queue->submit[tail] = cmd; // add to tail (end of queue)
+    tail++;
+    if (tail == queue->elements) {
+        tail = 0;
+    }
+    *(queue->submitdb) = tail; // set to tail
+    queue->sqtail = tail; // update tail so now it'll point to the element after (nothing until we submit a new command)
+}
+
+// submit a command an wait for completion
+static uint16_t nvme_awaitsubmitcmd(struct nvme_queue *queue, struct nvme_cmd cmd) {
+    uint16_t head = queue->cqhead;
+    uint16_t phase = queue->cqphase;
+    cmd.common.cid = queue->cmdid++;
+    nvme_submitcmd(queue, cmd);
+    uint16_t status = 0; 
+
+    while (true) {
+        status = queue->completion[queue->cqhead].status;
+        if ((status & 0x01) == phase) {
+            break;
+        }
+    }
+
+    status >>= 1;
+    ASSERT_MSG(!status, "nvme: cmd error %x", status);
+
+    head++;
+    if (head == queue->elements) {
+        head = 0;
+        queue->cqphase = !queue->cqphase; // flip phase
+    }
+
+    *(queue->completedb) = head;
+    queue->cqhead = head;
+    return status;
+}
+
 static ssize_t nvme_setqueuecount(struct nvme_device *ctrl, int count) {
     struct nvme_cmd cmd = { 0 };
     cmd.features.opcode = NVME_OPSETFT;
@@ -110,47 +402,6 @@ static ssize_t nvme_createqueues(struct nvme_device *ctrl, struct nvme_nsdevice 
     return 0;
 }
 
-// submit a command
-void nvme_submitcmd(struct nvme_queue *queue, struct nvme_cmd cmd) {
-    uint16_t tail = queue->sqtail; // tail of the submit queue
-    queue->submit[tail] = cmd; // add to tail (end of queue)
-    tail++;
-    if (tail == queue->elements) {
-        tail = 0;
-    }
-    *(queue->submitdb) = tail; // set to tail
-    queue->sqtail = tail; // update tail so now it'll point to the element after (nothing until we submit a new command)
-}
-
-// submit a command an wait for completion
-uint16_t nvme_awaitsubmitcmd(struct nvme_queue *queue, struct nvme_cmd cmd) {
-    uint16_t head = queue->cqhead;
-    uint16_t phase = queue->cqphase;
-    cmd.common.cid = queue->cmdid++;
-    nvme_submitcmd(queue, cmd);
-    uint16_t status = 0; 
-
-    while (true) {
-        status = queue->completion[queue->cqhead].status;
-        if ((status & 0x01) == phase) {
-            break;
-        }
-    }
-
-    status >>= 1;
-    ASSERT_MSG(!status, "nvme: cmd error %x", status);
-
-    head++;
-    if (head == queue->elements) {
-        head = 0;
-        queue->cqphase = !queue->cqphase; // flip phase
-    }
-
-    *(queue->completedb) = head;
-    queue->cqhead = head;
-    return status;
-}
-
 static ssize_t nvme_identify(struct nvme_device *ctrl, struct nvme_id *id) {
     uint64_t len = sizeof(struct nvme_id);
     struct nvme_cmd cmd = { 0 };
@@ -183,7 +434,7 @@ static ssize_t nvme_identify(struct nvme_device *ctrl, struct nvme_id *id) {
     return 0;
 }
 
-ssize_t nvme_nsid(struct nvme_nsdevice *ns, struct nvme_nsid *nsid) {
+static ssize_t nvme_nsid(struct nvme_nsdevice *ns, struct nvme_nsid *nsid) {
     struct nvme_cmd cmd = { 0 };
     cmd.identify.opcode = NVME_OPIDENTIFY;
     cmd.identify.nsid = ns->nsid; // differentiate from normal identify by passing name space id
@@ -283,7 +534,7 @@ notfound:
 }
 
 // read `count` bytes at `loc` into `buf`
-ssize_t nvme_read(struct resource *_this, struct f_description *description, void *buf, off_t loc, size_t count) {
+static ssize_t nvme_read(struct resource *_this, struct f_description *description, void *buf, off_t loc, size_t count) {
     (void)description;
     spinlock_acquire(&_this->lock);
     struct nvme_nsdevice *this = (struct nvme_nsdevice *)_this;
@@ -312,7 +563,7 @@ ssize_t nvme_read(struct resource *_this, struct f_description *description, voi
     return count;
 }
 
-ssize_t nvme_write(struct resource *_this, struct f_description *description, const void *buf, off_t loc, size_t count) {
+static ssize_t nvme_write(struct resource *_this, struct f_description *description, const void *buf, off_t loc, size_t count) {
     (void)description;
     spinlock_acquire(&_this->lock);
     struct nvme_nsdevice *this = (struct nvme_nsdevice *)_this;
